Move initializeData tests to new structure.
