+++ src/tests.cmake
@@ -56,7 +56,6 @@ target_sources(testprecice
     src/partition/tests/ReceivedPartitionTest.cpp
     src/partition/tests/fixtures.hpp
     src/precice/tests/DataContextTest.cpp
-    src/precice/tests/InitializeDataTests.cpp
     src/precice/tests/ParallelTests.cpp
     src/precice/tests/SerialTests.cpp
     src/precice/tests/TimeTests.cpp
