+++ tests/serial/initialize-data/testDataInitializationReadMapping.xml
