+++ tests/serial/initialize-data/testDataInitializationWriteMapping.xml
