+++ tests/serial/initialize-data/testExplicitWithDataInitialization.cpp
@@ -1,0 +1,75 @@
+#ifndef PRECICE_NO_MPI
+
+#include "testing/Testing.hpp"
+
+#include <precice/SolverInterface.hpp>
+#include <vector>
+
+using namespace precice;
+
+BOOST_AUTO_TEST_SUITE(PreciceTests)
+BOOST_AUTO_TEST_SUITE(Serial)
+BOOST_AUTO_TEST_SUITE(InitializeData)
+
+/**
+ * @brief The second solver initializes the data of the first.
+ *
+ * A mapping is employed for the second solver, i.e., at the end of
+ * initializeData(), the mapping needs to be invoked.
+ */
+BOOST_AUTO_TEST_CASE(testExplicitWithDataInitialization)
+{
+  PRECICE_TEST("SolverOne"_on(1_rank), "SolverTwo"_on(1_rank));
+
+  using Eigen::Vector3d;
+
+  SolverInterface cplInterface(context.name, context.config(), 0, 1);
+  if (context.isNamed("SolverOne")) {
+    int meshOneID = cplInterface.getMeshID("MeshOne");
+    cplInterface.setMeshVertex(meshOneID, Vector3d(1.0, 2.0, 3.0).data());
+    double maxDt      = cplInterface.initialize();
+    int    dataAID    = cplInterface.getDataID("DataOne", meshOneID);
+    int    dataBID    = cplInterface.getDataID("DataTwo", meshOneID);
+    double valueDataB = 0.0;
+    cplInterface.initializeData();
+    cplInterface.readScalarData(dataBID, 0, valueDataB);
+    BOOST_TEST(2.0 == valueDataB);
+    while (cplInterface.isCouplingOngoing()) {
+      Vector3d valueDataA(1.0, 1.0, 1.0);
+      cplInterface.writeVectorData(dataAID, 0, valueDataA.data());
+      maxDt = cplInterface.advance(maxDt);
+      cplInterface.readScalarData(dataBID, 0, valueDataB);
+      BOOST_TEST(2.5 == valueDataB);
+    }
+    cplInterface.finalize();
+  } else {
+    BOOST_TEST(context.isNamed("SolverTwo"));
+    int      meshTwoID = cplInterface.getMeshID("MeshTwo");
+    Vector3d pos       = Vector3d::Zero();
+    cplInterface.setMeshVertex(meshTwoID, pos.data());
+    double maxDt   = cplInterface.initialize();
+    int    dataAID = cplInterface.getDataID("DataOne", meshTwoID);
+    int    dataBID = cplInterface.getDataID("DataTwo", meshTwoID);
+    cplInterface.writeScalarData(dataBID, 0, 2.0);
+    //tell preCICE that data has been written and call initializeData
+    cplInterface.markActionFulfilled(precice::constants::actionWriteInitialData());
+    cplInterface.initializeData();
+    Vector3d valueDataA;
+    cplInterface.readVectorData(dataAID, 0, valueDataA.data());
+    Vector3d expected(1.0, 1.0, 1.0);
+    BOOST_TEST(valueDataA == expected);
+    while (cplInterface.isCouplingOngoing()) {
+      cplInterface.writeScalarData(dataBID, 0, 2.5);
+      maxDt = cplInterface.advance(maxDt);
+      cplInterface.readVectorData(dataAID, 0, valueDataA.data());
+      BOOST_TEST(valueDataA == expected);
+    }
+    cplInterface.finalize();
+  }
+}
+
+BOOST_AUTO_TEST_SUITE_END() // PreciceTests
+BOOST_AUTO_TEST_SUITE_END() // Serial
+BOOST_AUTO_TEST_SUITE_END() // InitializeData
+
+#endif // PRECICE_NO_MPI
