+++ tests/tests.cmake
@@ -7,4 +7,10 @@ target_sources(testprecice
     tests/serial/MultiCouplingFourSolvers2.cpp
     tests/serial/helpers.cpp
     tests/serial/helpers.hpp
+    tests/serial/initialize-data/helpers.cpp
+    tests/serial/initialize-data/helpers.hpp
+    tests/serial/initialize-data/testDataInitializationReadMapping.cpp
+    tests/serial/initialize-data/testDataInitializationWriteMapping.cpp
+    tests/serial/initialize-data/testExplicitWithDataInitialization.cpp
+    tests/serial/initialize-data/testImplicitWithDataInitialization.cpp
     )
